+++ packages/dartway_client/lib/src/client/calls.dart
@@ -303,6 +303,7 @@ extension on DwAppClient {
     DwHttpContract.contentTypeHeader: DwHttpContract.jsonContentType,
     DwHttpContract.protocolHeader: '$dwProtocolVersion',
     DwHttpContract.appVersionHeader: '$appVersion',
+    DwHttpContract.contractVersionHeader: ?protocol.contractVersion?.text,
     if (token != null)
       DwHttpContract.authorizationHeader:
           '${DwHttpContract.bearerPrefix}$token',